Cleanup chat adapter
