+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/ChatCompositeView.swift
@@ -7,11 +7,7 @@ import SwiftUI
 
 /// The Chat Composite View is a view component for a single chat thread
 public struct ChatCompositeView: View {
-
     let chatAdapter: ChatAdapter
-
-    let router: NavigationRouter
-    let logger: Logger
     let viewFactory: CompositeViewFactoryProtocol
     let isRightToLeft: Bool
 
@@ -20,24 +16,16 @@ public struct ChatCompositeView: View {
     ///    - chatAdapter: The required parameter to create a view component
     public init(with chatAdapter: ChatAdapter) {
         self.chatAdapter = chatAdapter
-
-        self.router = self.chatAdapter.navigationRouter!
-        self.logger = self.chatAdapter.logger
         self.viewFactory = self.chatAdapter.compositeViewFactory!
-
         let localizationProvider = self.chatAdapter.localizationProvider
         self.isRightToLeft = localizationProvider.isRightToLeft
-
     }
 
     /// The view body would be used to render the ChatCompositeView
     public var body: some View {
         VStack {
-            ContainerView(router: self.router,
-                          logger: self.logger,
-                          viewFactory: self.viewFactory,
+            ContainerView(viewFactory: self.viewFactory,
                           isRightToLeft: self.isRightToLeft)
         }
-
     }
 }
