+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/Factories/CompositeViewModelFactory.swift
@@ -26,7 +26,6 @@ protocol CompositeViewModelFactoryProtocol {
 class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
     private let logger: Logger
     private let localizationProvider: LocalizationProviderProtocol
-    private let accessibilityProvider: AccessibilityProviderProtocol
     private let messageRepositoryManager: MessageRepositoryManagerProtocol
     private let store: Store<ChatAppState, Action>
 
@@ -36,12 +35,10 @@ class CompositeViewModelFactory: CompositeViewModelFactoryProtocol {
     // - only skeleton code to show view, class not finalized yet
     init(logger: Logger,
          localizationProvider: LocalizationProviderProtocol,
-         accessibilityProvider: AccessibilityProviderProtocol,
          messageRepositoryManager: MessageRepositoryManagerProtocol,
          store: Store<ChatAppState, Action>) {
         self.logger = logger
         self.localizationProvider = localizationProvider
-        self.accessibilityProvider = accessibilityProvider
         self.messageRepositoryManager = messageRepositoryManager
         self.store = store
     }
