+++ AzureCommunicationUI/sdk/AzureCommunicationUIChat/Sources/Presentation/SwiftUI/Container/ContainerView.swift
@@ -6,26 +6,11 @@
 import SwiftUI
 
 struct ContainerView: View {
-
-    @ObservedObject var router: NavigationRouter
-
-    let logger: Logger
     let viewFactory: CompositeViewFactoryProtocol
     let isRightToLeft: Bool
 
     var body: some View {
-        Group {
-            switch router.currentView {
-            case .chatView:
-                chatView
-            }
-        }
+        viewFactory.makeChatView()
         .environment(\.layoutDirection, isRightToLeft ? .rightToLeft : .leftToRight)
     }
-
-    var chatView: ChatView {
-        logger.debug("Displaying view: chatView")
-        return viewFactory.makeChatView()
-    }
-
 }
